Refactor ReplCommands out of the manager
They are now independent files in alda.repl.commands
